........
  r64979 | thomas.heller | 2008-07-15 22:18:46 +0200 (Tue, 15 Jul 2008) | 1 line

  Fix test on 64-bit platforms.
........

+++ Lib/ctypes/test/test_pep3118.py
@@ -92,7 +92,7 @@ class Incomplete(Structure):
 class Complete(Structure):
     pass
 PComplete = POINTER(Complete)
-Complete._fields_ = [("a", c_int)]
+Complete._fields_ = [("a", c_long)]
 
 ################################################################
 #
